feat: added model evaluation feature in the training model script
