+++ .gitignore
@@ -77,6 +77,9 @@ target/
 # Jupyter Notebook
 .ipynb_checkpoints
 
+# pycache
+__pycache__
+
 # IPython
 profile_default/
 ipython_config.py
