+++ lib/modelbuilder.py
@@ -1,0 +1,29 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+class Model(nn.Module):
+    def __init__(self):
+        super(Model, self).__init__()
+        
+        self.conv1 = nn.Conv2d(3, 16, (3,3), padding=1)
+        self.conv2 = nn.Conv2d(16, 32, (3,3), padding=1)
+        self.conv3 = nn.Conv2d(32, 64, (3,3), padding=1)
+        self.pool = nn.MaxPool2d((2,2), 2)
+        
+        self.fc1 = nn.Linear(64*8*8, 128)
+        self.fc2 = nn.Linear(128, 3)
+        
+    def forward(self, x):
+        x = F.relu(self.conv1(x))
+        x = self.pool(x)
+        x = F.relu(self.conv2(x))
+        x = self.pool(x)
+        x = F.relu(self.conv3(x))
+        x = self.pool(x)
+        
+        x = x.view(-1, 64*8*8)
+        x = F.relu(self.fc1(x))
+        x = self.fc2(x)
+        
+        return x
